tests: move jest setup files to test folder
