+++ test/helpers/consoleSpies.ts
@@ -1,0 +1,3 @@
+export const logSpy = jest.spyOn(console, 'log')
+export const errorSpy = jest.spyOn(console, 'error')
+export const warnSpy = jest.spyOn(console, 'warn')
